requirements.txt
python-barcode
opencv-python
pyzbar
pillow
numpy
